Move styles to separated files
