+++ wchat/screens/ContactList/components/ContactCell/index.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import { AppRegistry, StyleSheet, Text, Image, View } from 'react-native';
+import { AppRegistry, Text, Image, View } from 'react-native';
+import styles from './styles.js'
 
 export default class ContactCell extends Component {
   render() {
@@ -17,43 +18,3 @@ export default class ContactCell extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-   flex: 1,
-  },
-  content: {
-   flex: 1,
-   flexDirection: 'row',
-   marginVertical: 12,
-  },
-  avatar: {
-    height: 44,
-    width: 44,
-    marginLeft: 16,
-    borderRadius: 22,
-    overflow: 'hidden',
-  },
-  info: {
-   flex: 1,
-   marginLeft: 10,
-   justifyContent: 'center',
-  },
-  name: {
-    fontSize: 16,
-    textAlign: 'left',
-    fontWeight: 'bold',
-    marginBottom: 4,
-  },
-  message: {
-    fontSize: 14,
-    textAlign: 'left',
-    color: '#999',
-  },
-  separator: {
-    height: 1,
-    backgroundColor: '#CCC',
-    marginRight: 10,
-    marginLeft: 70,
-  }
-})
