+++ wchat/screens/ContactList/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+   flex: 1,
+   paddingTop: 22
+  }
+})
+
+export default styles
